Ports TGMC timer cooldowns (now supporting signals) (#51388)
Added it to a random timer as proof of concept.

This is less performant than storing world.time in a variable, but it's a simple, quick, easy and dirty way to add a cooldown to something. Just create an ID and you need do no more besides using the macros.

Helps with clarity and adding cooldowns on the fly.