makes decal painter a basic tool (#51762)
* makes decal painter a basic tool

Makes the Decal Painter into a Basic Tool instead of a BEPIS unlock. It's very meh for a BEPIS unlock, especially for the Scientists that'll be using it. Though, it does have a use that some builders might like.

* Update maintenance_loot.dm

* decal painter value decrease

decal painter now contains much less mats since it is a basic tool

* decal painter no longer needs an empty ink cartridge to eject