Service Area Update (#51697)
About The Pull Request

The service area on meta did not have a department feeling to it, the bar was of on it's own side, the back of house was not connected.
The worst offender was that sad little room for the service TechFab.

This creates a better cohesive service department, all service areas have access to a proper service back hall with a proper delivery area. (OK the janitor is still left out but something had to give.)

In this setup the bartender has access to the kitchen, and cook has access to the bar.

change
Why It's Good For The Game

This should create better traffic flow and create more active area for the crew in the bar.
The bartender, cook, botanist, even clown can now work as a team. or kill the clown you know, do you.
Changelog

🆑
tweak: Service department area update on Meta
/🆑