Stacker to silo fix (#51852)
About The Pull Request

Currently when the stacker picks up an item it is stored in the stacker console and creats a log in the silo that 0 amount was added to the silo. This is in contrast to when you slap the stacker with a stack, it will go to the silo and not the stacker console.

This makes it so that all items that are picked up will go to the connected silo, just like slapping it would.

The gulag points are properly calculated and save to the point console for both methods.

Fixes #48587
Why It's Good For The Game

Stacker works as expected when it picks up stacks.
Changelog

🆑
fix: Stackers deposit to the silo when picking up stacks from the input side.
/🆑