Sweeps up some rarely used and unbalanced sources of research points (#52012)
About The Pull Request

Removes the research point gen modes from tesla coils and rad collectors.
Someone is gonna take some flak for removing these at some point, may as well be me.
Why It's Good For The Game

THE DESIGN DOC DICTACT-

Nah powergen should be valid on it's own, it shouldn't need to rely on other econs to make it worth the work. We can try that later, but not now.
Changelog

cl
del: Removed the research modes from tesla coils and radiation collectors.
/cl