tgui: Telecommunications UI (#52357)
* Telecommunication Machine UI Overhaul

This replaces the UI of the tcomm machines from the old html system to the new TGUI system.

* Telecommunication Machine UI Overhaul

This replaces the UI of the tcomm machines from the old html system to the new TGUI system.

* Updated TGUI.Bundle

* Housekeeping

Added final newline, shortened value-values, spread out boxes, renamed Tcomm to Telecomms, and made the backend use full numbers while the frontend used decimals.

Thanks for the feedback, y'all.