[MIRROR] Update Lavaland Syndie Atmosia (#143)
* Update Lavaland Syndie Atmosia (#52566)

* Update Lavaland Syndie Atmos

Expands the syndicate atmos department on the lavaland ruin to accommodate for more complex experiments, adds tier3 canisters and a stack of plasteel to make more, and a bluespace RPED to go along with the boxes of stock parts in storage so that the players can upgrade any machine they make.

I was feeling that the current syndicate base is a bit dated, and I'm planning on doing more changes to the ruin but for now this should suffice.

* update to replace hidden with visible pipe

* Update Lavaland Syndie Atmosia

Co-authored-by: Asha108 <68921923+Asha108@users.noreply.github.com>