[MIRROR] Pipe cleaner cable color fix (#270)
* Pipe cleaner cable color fix (#52759)

This PR fixes a bug where service cyborgs were unable to set a color of the pipe cleaner cable they possess due to not setting an actual color code for a cable colorization and a duplicit name variable.

Also adds an option for cancelling color choosing input as per maintainer suggestion.

* Pipe cleaner cable color fix

Co-authored-by: Arkatos1 <43862960+Arkatos1@users.noreply.github.com>