[MIRROR] You Can No Longer Place Force Fields On The Same Tile As Yourself (#281)
* You Can No Longer Place Force Fields On The Same Tile As Yourself (#52569)

About The Pull Request

Adds a check to the force field projector that prevents you from placing fields on the same tile as yourself.
Why It's Good For The Game

Some people have been using this recently in fights to become temporarily unclickable, which is kinda bad if you ask me.
Changelog

cl
tweak: you can no longer place force fields on the same tile as yourself to become god for 30 seconds
/cl

* You Can No Longer Place Force Fields On The Same Tile As Yourself

Co-authored-by: Paxilmaniac <57122098+Paxilmaniac@users.noreply.github.com>