[MIRROR] Prevent duplicate names for vents and scrubbers in the air alarm (#431)
* Prevent duplicate names for vents and scrubbers in the air alarm (#52398)

It is always a pain adding in a new vent or scrubber to an area and suddenly you have duplicate names.
This prevents duplicate names for vents and scrubbers in an area.
Why It's Good For The Game

Makes engineering's life better, duplicate names are bad.
Changelog

cl
fix: Prevent duplicate vent and scrubber names in the air alarm interface.
/cl

* Prevent duplicate names for vents and scrubbers in the air alarm

Co-authored-by: NightRed <nightred@gmail.com>