[MIRROR] Disposable Sentry Gun Deployment System (#495)
* Disposable Sentry Gun Deployment System (#52931)

Adds a new engineer only traitor item, the disposable sentry gun deployment system. It's a red toolbox that works and functions as a normal toolbox, until you attack it on harm intent with a wrench. Upon bashing it with a wrench, the toolbox will transform into a ballistic sentry gun, theoretically shooting at your enemies, and not you. The turret only has 100 health however, so it will not last long to anyone with any weapon that can deal enough damage.

* Disposable Sentry Gun Deployment System

Co-authored-by: Paxilmaniac <57122098+Paxilmaniac@users.noreply.github.com>