[MIRROR] AI core is now Abductor-proof. Part 2: This time they really can't get in, promise. (#556)
* Merge pull request #53005 from Wayland-Smithy/ayy_send_self_AI

Duplicates the "turret_protected/ai" area type check found on the alien abductors camera console Send To command so it also applies to the Send Self command.

Getting snatched by abductors as AI Post #52718 is bad.

This is not an IDED because it was caused by a glitch. I was playing as AI and found myself carded by some ayylines. It turns out while they can not send their agent to do the deed they were more than capable of sending self as I suspected and confirmed with some local testing.

* AI core is now Abductor-proof. Part 2: This time they really can't get in, promise.

Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>