[MIRROR] Adds an admin-only simple teambuilding machine. (#557)
* Merge pull request #53147 from ArcaneMusic/teambuilding

We needed an admin tool for the "Cool Map" PR, to be able to set up a quick blue and red team, in order to drive conflict.
This adds a machine that when walked over, will assign anyone who walks over the teambuilder with:

    A color, assigned via team_color
    A radio channel, via team_radio.
    One, or the other, or both can be assigned and it will enforce that, color won't override an existing one but in most cases players won't override their own color.

Big admin use for events, plus all it takes is 2 VVs and you're good to go.

* Adds an admin-only simple teambuilding machine.

Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>