[MIRROR] You can buckle yourself or somebody else into a chair, if you're next to the chair. (#593)
* You can buckle yourself or somebody else into a chair, if you're next to the chair. (#53255)

You can now buckle yourself, as well as others, into a chair as long as you (and the other person, if you're not buckling yourself) are Adjacent to said chair, and being on the same tile as the chair is no longer required.

No more god-awful pixelhunt to accomplish a very simple task.

(You can buckle yourself and others into chairs that you're right next to, instead of needing to be on top of the chair to do it.)

* You can buckle yourself or somebody else into a chair, if you're next to the chair.

Co-authored-by: uomo <51800976+uomo91@users.noreply.github.com>