[MIRROR] Add warning about https to the external rsc config. (#631)
* Add warning about https to the external rsc config (#53367)

today in hostchat we discovered that most of the servers who tried to 
do cdn the .rsc, had unknown issues because they had used https:// 
urls. Byond can't into https.

like 5 servers had failed to get this to work because of that pitfall, 
so im gonna add a warning to the config.

* Add warning about https to the external rsc config.

Co-authored-by: Kyle Spier-Swenson <kyleshome@gmail.com>