[MIRROR] Makes the stomach important part of eating (#633)
* Makes the stomach important part of eating (#53228)

This puts food you eat in to the stomach.
If you do not have a stomach you can not eat, if you try to swallow anything you will spew it out.

Changes to surgeries:

    Stomach pump can now be done on the living, and pushes out the reagents in the stomach.
    Blood filter is a new tool and cleans the reagents out of the body.

This also makes it so that reagents that do not metabolize can accumulate in the stomach reducing how much food you can eat.

* Makes the stomach important part of eating

Co-authored-by: NightRed <nightred@gmail.com>