[MIRROR] Allows for hostile mobs to perform charge attacks, and adds a giant tackling lobster.  (#651)
* Allows for hostile mobs to perform charge attacks, and adds a giant tackling lobster.  (#52692)

Example video: https://youtu.be/EAVuPtJ_kFM

This adds a new mob, the lobstrosity. It's a upper-caves level icemoon mob, created as a genetic combination of a lobster, a crab, and a shrimp, let loose on the horrific Alaskan hellscape that is the icemoon. It's best feature, however, is that it makes use of a new feature of simplemobs, charging. Any hostile simplemob can be configured and customized in a uniform way in order to become a charger, giving it a basic leap attack, that will disable the user if they find themselves crashing into an unpassable turf, but otherwise lands a knockdown on the attacking target. Compensate of this, charging mobs will also gain a very noticable tell, where they shake a moment before they actually perform the charge to better signify that they're rearing up for a melee charge.
I'm not exactly sold on using Shake() for the visual tell, but I'm not quite sold on a visual to showcase that it's happening, so I'm looking for feedback on that specifically.
Additionally, there is a bit of weird movement that I've yet to really get ironed out first but for the most part it's completely functional.

After some feedback, there's now a reddish, lavaland-colored lobster, as well as an arctic themed, Alaskan lobster

So on a massive tangential leap of logic, one thing ActionNinja brought up in a recent discord rant was that there are so many mechanical parts of SS13 combat that completely lack a tell whatsoever, or the tell is so quick you can blink and you'll miss it in a heartbeat. So I turned to the greatest enemy attack feedback mechanic that any game has ever seen: the Legend of Zelda Bull.

...also I've been looking for a tasteful way to get this sprite in-game for awhile now but wanted to really iron out a unique mechanic to go alongside it that would make it feel a little less bullshit to fight against.

ArcaneMusic + Kryson on Sprites

* Allows for hostile mobs to perform charge attacks, and adds a giant tackling lobster.

Co-authored-by: ArcaneMusic <41715314+ArcaneMusic@users.noreply.github.com>