[MIRROR] Wild Mutation crops now carry over traits from the parent plant. (#663)
* Wild Mutation crops now carry over traits from the parent plant. (#53413)

This results in wild mutation carrying over possible traits from parent to child, in a way that will also be displayed by the plant analyzer.

This means if you gene splice a parent plant and it wild mutates into a different species, the wild mutation will take the trait genes of the parent, with none of the reagents as reagents are specific to the species in most cases, but can be bred back in via cross-pollination.

* Wild Mutation crops now carry over traits from the parent plant.

Co-authored-by: ArcaneMusic <41715314+ArcaneMusic@users.noreply.github.com>