[MIRROR] Fix Cargo tablet bounty programs not being able to print the bounty form (#668)
* Fix Cargo tablet bounty programs not being able to print the bounty form (#53365)

Fixes the printer cool down system that started the printer cool down when you press any button??? meaning that it would always fail since the cool down was reset after the button was clicked. The cool down is now only reset until AFTER the print operation.
Also adds a message that lets you know if the print cool down is not over yet, so it doesn't fail without feedback.

* Fix Cargo tablet bounty programs not being able to print the bounty form

Co-authored-by: Jack7D1 <59462654+Jack7D1@users.noreply.github.com>