[MIRROR] Fixes the arena shuttle (#691)
* Fixes the arena shuttle (#53458)

This splits the actual arena of the arena shuttle into its own map file that is loaded onto it's own z-level. The old method of including a non-shuttle area in the shuttle map file doesn't work with the current shuttle loader. It also fixes shuttles being loaded twice when purchased which created problems with landmarks being left behind in the reserved z-level.

* Fixes the arena shuttle

Co-authored-by: YPOQ <30683121+YPOQ@users.noreply.github.com>