[MIRROR] You can now tap TCG cards. (#707)
* You can now tap TCG cards. (#53459)

Basically the last groundwork TCG PR we need before we can sew the rules into the game, you can now use TCG cards on the ground with a new radial with an empty hand to either tap, flip, or pick up individual cards. This will also mean that using a deck on a set of individual cards on the ground is now the fastest and most efficient way to pick up cards now.

 Author's note: Opponent's card was not "deleted for effect", it was sent to the Shadow Realm, for all eternity.

* You can now tap TCG cards.

Co-authored-by: ArcaneMusic <41715314+ArcaneMusic@users.noreply.github.com>