[MIRROR] Adds native chilly peppers to icebox flora. (#745)
* Adds native chilly peppers to icebox flora. (#53544)

Now, icebox has a native plant that can be picked on the surface level, where the sunlight reaches highest, and rarely seen underground, and provides you chilly peppers. Traditionally a chili pepper mutation, the theme was close enough that I felt it'd be a good fit for the map, and helps break up the monotony of the map being ice, rock, and more ice with some harvest able flora.

* Adds native chilly peppers to icebox flora.

Co-authored-by: ArcaneMusic <41715314+ArcaneMusic@users.noreply.github.com>