[MIRROR] Adds rollerskates, wheely can do tricks (#751)
* Adds rollerskates, wheely can do tricks (#53497)

This turns wheelys into a skateboard subtype so they are no longer a copy paste of old skateboard code. They have lower stability compared to skateboards.
I've also made a whole bunch of small adjustments in the process to make skateboard code cleaner and more modular, such as turning the improvised skateboard to its own subtype so that construction steps don't need to be overwritten for all subtypes

Adds two subtypes of wheelys:
-Rollerskates: more stable than wheelys but slow you down when the wheels are retracted
-Skishoes: For skiing! Only work on snow, obviously

I wanted to add some fun alternatives to skateboards with their own drawbacks, and the ski shoes were an opportunity to do something fun for icemoon maps

* Adds rollerskates, wheely can do tricks

Co-authored-by: Mickyan <38563876+Mickyan@users.noreply.github.com>