[MIRROR] Fixes some issues with sign language & makes some things more clear (#783)
* Fixes some issues with sign language & makes some things more clear (#53610)

Changes the name of the recipe for the gloves to make them more clear
Adds a note in the tongue file to make sure it's clear you SHOULDN'T SEE THAT TONGUE.
Adds a unique chat message for mimes who use sign language, to make it clearer you can't cheese as a mime
Adds a unique slurring verb for drunkards
Removes the ability for sign languagers to use megaphones, which I forgot existed in the game. 
Punctuation removal now replaces all punctuation used with periods rather than with blank space, to reduce run-on sentences.

* Fixes some issues with sign language & makes some things more clear

Co-authored-by: Wallemations <66052067+Wallemations@users.noreply.github.com>