[MIRROR] Fix EMP-shielded camera runtime on destruction (#837)
* Fix EMP-shielded camera runtime on destruction (#53736)

Cameras delete their assemblies on Destroy(). The assembly calls RemoveComponent() on the Camera's emp component. This nulls the emp_component's parent. The camera itself then attempts to QDEL_NULL its emp component, which leads to the above image of a SEND_SIGNAL to a parent that's null.

Since the assembly manages it all for us, including calling the appropriate emp removal proc, we can remove QDEL_NULL in the camera Destroy() and instead QDEL_NULL the emp component where it once was simply RemoveComponent()'d.

* Fix EMP-shielded camera runtime on destruction

Co-authored-by: Timberpoes <silent_insomnia_pp@hotmail.co.uk>