[MIRROR] Fixes processing of deleted objects in SSair (#828)
* Fixes processing of deleted objects in SSair (#53735)

qdel'd objects could still process in SSair because SSair utilises a cache whenever it has to resume processing runs from a partially completed state from running out of processing time.

Of all the things that processed on SSair, only one thing actually took care to remove itself from the cache as well on deletion.

This is an important subsystem and the processing lists should not be public. Objects don't need to know how SSair works, they just call the proc to add when they start processing and call the proc to remove when they finish.

Thanks to @ LemonInTheDark and @ willox for spending a lot of time helping me track down a proper fix to this issue.

* Fixes processing of deleted objects in SSair

Co-authored-by: Timberpoes <silent_insomnia_pp@hotmail.co.uk>