[MIRROR] Pubbystation mapping fixes (#873)
* Pubbystation mapping fixes (#53780)

Adds the QM spawn in their office overlooking the cargo bay and fixes cargo techs not being able to access the delivery office. I didn't change the mining office door, because as far as I know only the QM and miners can access mining on most stations.
EDIT: I thought that was a cargo tech in the picture attached to the access issue. Fixed the mining door to include QM access.

As I was adding the QM spawn I noticed that the miner spawns and cargo tech spawns were missing as well, so I added them in their usual spots.

* Pubbystation mapping fixes

Co-authored-by: tattlemothe <66640614+dragomagol@users.noreply.github.com>