[MIRROR] Excessive spinning makes you woozy (#878)
* Excessive spinning makes you woozy (#53805)

When attempting to spin while it's on cooldown carbons get a bit of dizziness and confusion in increments, and once it reaches a certain point they no longer get more confused, instead puking without purging their reagents.
This doesn't affect simplemobs, silicons, or using fists of the northstar to beyblade as a tot.

* Excessive spinning makes you woozy

Co-authored-by: ArcaneDefence <51932756+ArcaneDefence@users.noreply.github.com>