[MIRROR] Healium is too powerful we need to nerf it! (#908)
* Healium is too powerful we need to nerf it! (#53766)

Healium was too stronk, nerfed it to behave like a good gas.
Added coughing and euphoria mood if breathed in small quantities to alert the player.
Sleep if it is 3 moles in the air for 3 to 5 seconds
Sleep you if is more than 6 moles and it heals you while knocked down.
Lowered the amount of healing done

Edit: It doesnt knock you out, it makes you sleep now, since the knockdown effect was so strong that people wouldn't wake up

* Healium is too powerful we need to nerf it!

Co-authored-by: Ghilker <42839747+Ghilker@users.noreply.github.com>