[MIRROR] adds headpatting (#924)
* Merge pull request #53855 from imsxz/twaticus

About The Pull Request
i made a PR just like this like a year or 2 ago and it got closed under dubious circumstances, especially since it had more upvotes than downvotes. this is a feature that the community wants. it is a 5 line change PR, not much to note on it, very simple.

Why It's Good For The Game
/tg/ is lacking in roleplay tailored mechanics and instead relies on the good graces of the players to interact politely with each other.

Changelog
🆑 imsxz
add: you can now pat carbon mobs on the head.
/🆑

* adds headpatting

Co-authored-by: skoglol <33292112+kriskog@users.noreply.github.com>