[MIRROR] Prevent inventory HUDs from overlapping multiple overlays (#1042)
* Prevent inventory HUDs from overlapping multiple overlays (#53974)

Currently, when you drag your mouse off an inventory button while holding it down, it'll leave a permanent ghost image (#53950). This is because MouseExited is not being called properly. However, I couldn't reproduce this in a non-SS13 BYOND project, which makes it difficult for me to fix that.

This is a band-aid fix. It'll leave a ghost image, but then the next time you hover over and exit, it'll remove it, making it much less annoying.

* Prevent inventory HUDs from overlapping multiple overlays

Co-authored-by: Jared-Fogle <35135081+Jared-Fogle@users.noreply.github.com>