[MIRROR] Nice, Corgi Pinning Action for the whole family. (#1077)
* Nice, Corgi Pinning Action for the whole family. (#53913)

Adds pin the tail on the corgi, a basic party game last seen at the summer ball, that I had lying around on one of my branches since then. The corgi is a poster that functions like most other posters, but has an interaction with the fake "Corgi tail" object, which can be pinned onto the poster to visually place the tail onto the poster.

Are you a bad enough dude to pin the tail on the corgi blindfolded?

Fun, cheap entertainment. Might make for a cheap thrill, or help set the scene of a birthday party. Works as a very player driven basic activity.

Also, the tail pins are harmless embeds. I'll leave that here for you to think about.

* Nice, Corgi Pinning Action for the whole family.

Co-authored-by: ArcaneMusic <41715314+ArcaneMusic@users.noreply.github.com>